requirements.txt
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Deprecated==1.2.14
Flask==3.0.3
Flask-Limiter==3.6.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
limits==3.11.0
mariadb==1.1.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.0
pillow==10.3.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
rich==13.7.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
